requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes --no-editable --no-emit-project --output-file=requirements.txt
aiofiles==24.1.0
    # via gradio
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   google-genai
    #   gradio
    #   httpx
    #   starlette
audioop-lts==0.2.2
    # via gradio
brotli==1.1.0
    # via gradio
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1 ; sys_platform != 'emscripten'
    # via
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
fastapi==0.116.1
    # via gradio
ffmpy==0.6.1
    # via gradio
filelock==3.19.1
    # via huggingface-hub
fsspec==2025.9.0
    # via
    #   gradio-client
    #   huggingface-hub
google-api-core==2.25.1
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-genai
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via google-cloud-logging
google-cloud-logging==3.12.1
    # via combo-banana
google-genai==1.33.0
    # via combo-banana
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
gradio==5.44.1
    # via combo-banana
gradio-client==1.12.1
    # via gradio
groovy==0.1.2
    # via gradio
grpc-google-iam-v1==0.14.2
    # via google-cloud-logging
grpcio==1.74.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.74.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.9 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   gradio
    #   gradio-client
    #   safehttpx
huggingface-hub==0.34.4
    # via
    #   gradio
    #   gradio-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jinja2==3.1.6
    # via gradio
markdown-it-py==4.0.0 ; sys_platform != 'emscripten'
    # via rich
markupsafe==3.0.2
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2 ; sys_platform != 'emscripten'
    # via markdown-it-py
numpy==2.3.2
    # via
    #   gradio
    #   pandas
opentelemetry-api==1.36.0
    # via google-cloud-logging
orjson==3.11.3
    # via gradio
packaging==25.0
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
pandas==2.3.2
    # via gradio
pillow==11.3.0
    # via
    #   combo-banana
    #   gradio
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
protobuf==6.32.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   combo-banana
    #   fastapi
    #   google-genai
    #   gradio
pydantic-core==2.33.2
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.2 ; sys_platform != 'emscripten'
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-multipart==0.0.20
    # via gradio
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   gradio
    #   huggingface-hub
requests==2.32.5
    # via
    #   google-api-core
    #   google-genai
    #   huggingface-hub
rich==14.1.0 ; sys_platform != 'emscripten'
    # via typer
rsa==4.9.1
    # via google-auth
ruff==0.12.12 ; sys_platform != 'emscripten'
    # via gradio
safehttpx==0.1.6
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4 ; sys_platform != 'emscripten'
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via
    #   fastapi
    #   gradio
tenacity==9.1.2
    # via google-genai
tomlkit==0.13.3
    # via gradio
tqdm==4.67.1
    # via huggingface-hub
typer==0.17.4 ; sys_platform != 'emscripten'
    # via gradio
typing-extensions==4.15.0
    # via
    #   fastapi
    #   google-genai
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   gradio
    #   requests
uvicorn==0.35.0 ; sys_platform != 'emscripten'
    # via gradio
websockets==15.0.1
    # via
    #   google-genai
    #   gradio-client
zipp==3.23.0
    # via importlib-metadata
